LAST COMMIT COMPATIBLE WITH D2JAM TECH DEMO
CHANGES:
- fixed window aspect ratio
- cleaned project and updated its settings

REMARK: Perhaps one could use in the future "keep_height" instead of "keep".
This would allow for usage on extreme wide screen devices.
If using the parallax background, there will be always art to display horizontally.
Due to the fixed height in this mode, there will be no issues with vertical gaps.
However, this approach makes seamless transition between different visual scenes more different, as the visibility of the transition due to e.g. loading is dependent on the screen size and therefor hard to test.