cleaned project/visuals and added railcar horn sound effects
CHANGES:
- cleaned railcar directory structure
- added railcar horn sound effects
- added panel container to credits text

BUG:
Railcar horn can be triggered multiple times in a row, leading to a very loud playback due to Godot Meta Player stacking.
To fix this, some blocking based upon whether the selected horn sound is playing is required. For that to be possible, there have to changes made to the Godot Meta Player addon.